+++ Client/Core/Commands/MiscHandler.cs
@@ -5,6 +5,7 @@ using System.Net;
 using System.Threading;
 using System.Windows.Forms;
 using xClient.Core.Helper;
+using xClient.Core.Networking;
 
 namespace xClient.Core.Commands
 {