+++ Client/Core/ReverseProxy/Packets/ReverseProxyConnectResponse.cs
@@ -1,5 +1,6 @@
 ﻿using ProtoBuf;
 using System;
+using xClient.Core.Networking;
 using xClient.Core.Packets;
 
 namespace xClient.Core.ReverseProxy.Packets