+++ Server/Forms/FrmTaskManager.cs
@@ -2,6 +2,7 @@
 using System.Windows.Forms;
 using xServer.Core;
 using xServer.Core.Misc;
+using xServer.Core.Networking;
 
 namespace xServer.Forms
 {